implement faster port-broker
It's still not super fast, but it is a big improvement.  I'm not sure
how much faster it can go, because the core problem is that the port
interface is *fundamentally* byte-based, and every operation has to go
through bytes, whereas everything I want to do is character-based.  So
at every turn I have to translate bytes to characters and back several
times.