get parse-direct port location setting *actually* working
This was very difficult to debug, because it worked perfectly when I
tested the port-broker file, but things were just crazy testing
scheduler.  It turns out that in the port-broker file I enabled
`port-count-lines!` manually in the tests, which I didn't do in the
scheduler.  If you don't call `port-count-lines!`, the custom location
function is never called, and some stupid counter just counts bytes
read or something!  Setting counting on each port as it is created in
the port broker fixes the problem...