readtable: fix promise-no-left-recursion
and other issues with left-recursive-nonterminating parsers.  In
particular this solves the issue that caused the bnf implementation to
have infinite loops if all parsers were labelled as
left-recursive-nonterminating -- the readtable implementation was
promising that they were NOT left recursive, and so didn't capture
continuations.  That was a serious oversight when I added
left-recursive-nonterminating as an option...